extract code to set up app urls
